swiss_stats/requirements.txt
swisslandstats-geopy
geopandas
pandas
numpy
openpyxl
pyarrow
